requirements.txt
bootstrap4==0.1.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-cart==1.0.4
django-crispy-forms==1.7.2
django-filter==2.1.0
django-tinymce==2.8.0
django-widget-tweaks==1.4.3
idna==2.8
numpy==1.16.1
Pillow==5.4.1
pytz==2018.7
requests==2.21.0
stripe==2.21.0
urllib3==1.24.1
virtualenv==16.1.0
